requirements.txt
alabaster==0.7.6
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
biopython==1.66
certifi==2015.11.20
decorator==4.0.4
docutils==0.12
functools32==3.2.3.post2
ipykernel==4.2.0
ipyparallel==4.1.0
ipython==4.0.1
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
MarkupSafe==0.23
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
nose==1.3.7
notebook==4.0.6
numpydoc==0.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
Pygments==2.0.2
pytz==2015.7
pyzmq==15.1.0
qtconsole==4.1.1
requests==2.8.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.9
terminado==0.5
testpath==0.2
tornado==4.3
traitlets==4.0.0
